Add the renderProjects.ts helper to src/helpers/
The renderProjects.ts helper exports by default the renderProjectCards function. This function first generates HTML cards for each project of the array passed in the second argument and appends it to the element passed as the first argument. It uses helper functions like generateProjectCard and generateProjectTag to successfully render the projects cards to the UI.