Created the navbar.ts helper file including new methods
The navbar.ts file includes three (3) methods for implementing a navigation menu for the page's mobile version. The checkViewport function takes a width value as a parameter and return true if the width given is greater or equal to the viewport's width. The showNavbar and hideNavbar methods show and hide the navigation bar respectively by adding or removing the nav-visible and nav-drop classes from the navbar component.